Decimal stuff

+++ MoebiusEngine/src/main/java/civitas/celestis/math/matrix/FloatMatrix.java
@@ -497,7 +497,7 @@ public class FloatMatrix implements Matrix<Float, FloatMatrix> {
 
         for (int r = 0; r < rows; r++) {
             for (int c = 0; c < columns; c++) {
-                result.values[r][c] = (float) f.apply(values[r][c]);
+                result.values[r][c] = f.apply(values[r][c]);
             }
         }
 
@@ -682,7 +682,7 @@ public class FloatMatrix implements Matrix<Float, FloatMatrix> {
      *
      * @param s String to parse
      * @return Matrix parsed from given string
-     * @throws NumberFormatException    When the string is not parsable to a matrix
+     * @throws NumberFormatException When the string is not parsable to a matrix
      */
     @Nonnull
     public static FloatMatrix parseMatrix(@Nonnull String s) throws NumberFormatException {
